pyproject.toml
[project]
name = "digital-shop-backend"
version = "0.1.0"
description = "Digital product backend API"
authors = [
    {name = "Ayi",email = "erictim1217@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",

    "firebase-admin (>=6.8.0,<7.0.0)",
    "google-auth (>=2.40.2,<3.0.0)",
    "google-cloud-firestore (>=2.20.2,<3.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",

    "stripe (>=12.1.0,<13.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
